move text based ota-metafiles parsing logic from app.proto back to ota-metadata
